+++ hat/assets/js/apps/Iaso/components/forms/EditableTextFields.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import InputComponent from '../../../../components/forms/InputComponent';
+import InputComponent from './InputComponent';
 
 const EditableTextFields = ({ fields }) => {
     return fields.map(field => (
