+++ hat/assets/js/apps/Iaso/components/forms/UneditableFields.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import InputComponent from '../../../../components/forms/InputComponent';
+import InputComponent from './InputComponent';
 
 const UneditableFields = ({ fields }) => {
     return fields.map(field => (
